Initial setup of PQ
This basically just lets me get into the map..
